requirements.txt
Brotli @ file:///C:/ci/brotli-split_1602517705222/work
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255207413/work
click==7.1.2
cryptography @ file:///C:/ci/cryptography_1607639129468/work
dash @ file:///tmp/build/80754af9/dash_1607621005479/work
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
Flask==1.1.2
Flask-Compress @ file:///tmp/build/80754af9/flask-compress_1605067594328/work
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mysql-connector-python==8.0.18
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
pandas @ file:///C:/ci/pandas_1609008089615/work
plotly @ file:///tmp/build/80754af9/plotly_1610403957963/work
protobuf==3.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyMySQL @ file:///C:/ci/pymysql_1610288240654/work
python-dateutil==2.8.1
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1603979006023/work
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
retrying==1.3.3
scipy @ file:///C:/ci/scipy_1597686737426/work
six @ file:///C:/ci/six_1605205426665/work
SQLAlchemy @ file:///C:/ci/sqlalchemy_1608243783929/work
Werkzeug==1.0.1
wincertstore==0.2
